+++ src/main/java/group/one/dronefeeder/service/DroneService.java
@@ -1,12 +1,12 @@
 package group.one.dronefeeder.service;
 
+import java.util.List;
+import java.util.Optional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import group.one.dronefeeder.exception.DroneExistenteException;
 import group.one.dronefeeder.model.Drone;
 import group.one.dronefeeder.repository.DroneRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import java.util.List;
-import java.util.Optional;
 
 /**
  * Class DroneService.
@@ -32,7 +32,7 @@ public class DroneService {
     return droneJpaRepository.findAll();
   }
 
-  public Optional<Drone> findDroneById(Integer id) {
+  public Optional<Drone> findDroneById(Long id) {
     return droneJpaRepository.findById(id);
   }
 }
